[AM] Check if record is excluded or restricted - offenderId as long not in

+++ app/interfaces/OffenderApi.java
@@ -5,5 +5,5 @@ import java.util.concurrent.CompletionStage;
 public interface OffenderApi {
     CompletionStage<String> logon(String username);
 
-    CompletionStage<Boolean> canAccess(String bearerToken, int offenderId);
+    CompletionStage<Boolean> canAccess(String bearerToken, long offenderId);
 }
